pyproject.toml
[project]
name = "dev-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "google-genai>=1.49.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
